requirements.txt
black==22.8.0
Cerberus==1.3.4
click==8.1.3
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
PyYAML==6.0
tomli==2.0.1
